move chat room actions to chat component
